bot/requirements.txt
aiogram
python-dotenv
yoyo-migrations
asyncpg